+++ src/flask/config.py
@@ -11,10 +11,10 @@
 import errno
 import os
 import types
+import warnings
 
 from werkzeug.utils import import_string
 
-from . import json
 from ._compat import iteritems
 from ._compat import string_types
 
@@ -176,30 +176,62 @@ class Config(dict):
             if key.isupper():
                 self[key] = getattr(obj, key)
 
-    def from_json(self, filename, silent=False):
-        """Updates the values in the config from a JSON file. This function
-        behaves as if the JSON object was a dictionary and passed to the
-        :meth:`from_mapping` function.
+    def from_file(self, filename, load, silent=False):
+        """Update the values in the config from a file that is loaded
+        using the ``load`` parameter. The loaded data is passed to the
+        :meth:`from_mapping` method.
 
-        :param filename: the filename of the JSON file.  This can either be an
-                         absolute filename or a filename relative to the
-                         root path.
-        :param silent: set to ``True`` if you want silent failure for missing
-                       files.
+        .. code-block:: python
 
-        .. versionadded:: 0.11
+            import toml
+            app.config.from_file("config.toml", load=toml.load)
+
+        :param filename: The path to the data file. This can be an
+            absolute path or relative to the config root path.
+        :param load: A callable that takes a file handle and returns a
+            mapping of loaded data from the file.
+        :type load: ``Callable[[Reader], Mapping]`` where ``Reader``
+            implements a ``read`` method.
+        :param silent: Ignore the file if it doesn't exist.
+
+        .. versionadded:: 1.2
         """
         filename = os.path.join(self.root_path, filename)
 
         try:
-            with open(filename) as json_file:
-                obj = json.loads(json_file.read())
+            with open(filename) as f:
+                obj = load(f)
         except IOError as e:
             if silent and e.errno in (errno.ENOENT, errno.EISDIR):
                 return False
+
             e.strerror = "Unable to load configuration file (%s)" % e.strerror
             raise
+
         return self.from_mapping(obj)
+
+    def from_json(self, filename, silent=False):
+        """Update the values in the config from a JSON file. The loaded
+        data is passed to the :meth:`from_mapping` method.
+
+        :param filename: The path to the JSON file. This can be an
+            absolute path or relative to the config root path.
+        :param silent: Ignore the file if it doesn't exist.
+
+        .. deprecated:: 1.2
+            Use :meth:`from_file` with :meth:`json.load` instead.
+
+        .. versionadded:: 0.11
+        """
+        warnings.warn(
+            "'from_json' is deprecated and will be removed in 2.0."
+            " Use 'from_file(filename, load=json.load)' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from .json import load
+
+        return self.from_file(filename, load, silent=silent)
 
     def from_mapping(self, *mapping, **kwargs):
         """Updates the config like :meth:`update` ignoring items with non-upper